requirements.txt
django>=1.9
git+http://github.com/dashdrum/dashdrum-django-snips.git#egg=dashdrum-django-snips
django-annoying
django-braces
whitenoise
djangorestframework
django-debug-toolbar
gunicorn
git+http://github.com/dashdrum/django-heartbeat.git#egg=django-heartbeat
